refactor: use a custom hook dedicated to ISOstring conversion
