+++ src/ui/view/dataverse/component/generalMetadata/generalMetadata.tsx
@@ -1,14 +1,11 @@
-import * as E from 'fp-ts/Either'
 import type { FC } from 'react'
 import { useCallback, memo } from 'react'
 import { useTranslation } from 'react-i18next'
 import type { IconName } from '@/ui/component/icon/icon'
 import { Icon } from '@/ui/component/icon/icon'
-import { activeLanguageWithDefault } from '@/ui/languages/languages'
-import { convertToLocalizedYearIfISODateTime } from '@/util/isoDateTime/isoDateTime'
 import './generalMetadata.scss'
 import './i18n/index'
-import { pipe } from 'fp-ts/lib/function'
+import { useLocalizedYearIfISODateTime } from '@/ui/hook/useLocalizedYearIfISODateTime'
 
 export type MetadataProperty = {
   property: string
@@ -37,27 +34,8 @@ type GeneralMetadataListProps = {
 
 export const GeneralMetadataList: FC<GeneralMetadataListProps> = ({ metadata }) => {
   const { t } = useTranslation('generalMetadata')
-  const { lng } = activeLanguageWithDefault()
-
-  const handleLocalizedYearOrError = useCallback(
-    (fallback: string) =>
-      E.fold(
-        (error: Error) => {
-          console.error(error.message)
-          return fallback
-        },
-        (localizedYear: string) => localizedYear
-      ),
-    []
-  )
-
-  const convertValueToLocalizedYearIfISODateTime = useCallback(
-    (value: string) =>
-      pipe(
-        convertToLocalizedYearIfISODateTime(value, lng),
-        handleLocalizedYearOrError(t('generalMetadata.invalidDate'))
-      ),
-    [handleLocalizedYearOrError, lng, t]
+  const convertValueToLocalizedYearIfISODateTime = useLocalizedYearIfISODateTime(
+    t('generalMetadata.invalidDate')
   )
 
   const displayValue = useCallback(
